requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
prompt-toolkit==1.0.14
pycodestyle==2.5.0
Pygments==2.6.1
PyInquirer==1.0.3
pylint==2.4.4
python-dotenv==0.10.3
regex==2020.4.4
requests==2.22.0
requests-file==1.4.3
six==1.13.0
tldextract==2.2.2
urllib3==1.25.7
wcwidth==0.1.9
wrapt==1.11.2
